ai-engine/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0